+++ GameObjectLib/src/Components/Entity.cpp
@@ -13,7 +13,7 @@
 Entity::Entity() : healthPoint(100), maxHealthPoint(100), damage(10), attackSpeed(1.0f), maxSpeed(300.f), range(1), speed(300.f) {}
 Entity::Entity(const int& _hp, const int& _damage, const float& _speed, const float& _attackSpeed, const float& _range) : healthPoint(_hp), maxHealthPoint(_hp), damage(_damage), speed(_speed), maxSpeed(_speed + 100.f), attackSpeed(_attackSpeed), range(_range) {}
 
-void Entity::TakeDamage(const int& _damage)
+void Entity::TakeDamage(const float& _damage)
 {
 	if (!isInvicible) {
